removed warning.
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@4948 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0

+++ dune/fem/space/dgspace/dgbasefunctions.hh
@@ -535,7 +535,7 @@ namespace Dune
     {
       const int maxBaseFunctions = DGNumberOfBaseFunctions< polOrd, dimDomain >::numBaseFunctions;
 
-      assert( geometry().dim() == dimDomain );
+      assert( (int) geometry().dim() == dimDomain );
 
       if( dimDomain == 1 )
         return BaseFunctionCreator< 1, GeometryType::cube, maxBaseFunctions, 0 >::create( i );
